Make changes needed to get a runnable test program
This commit incorporates the changes needed to eliminate the undefined
references, and to fix all of the code that caused warning messages.  It
also adds a build script file that will build the object library then
create the executable test program tstr.

The current program has a known error in one of the imported modules that
causes a segmentation fault on exit after the return from main(), which is
one driving reason for this repository.  Another reason for the repository
is to check out the github push capabilities.