Segmentation fault problem fixed - global wrong init order
The global variable str_utf8 plbck_strg is instantiated before the ivs
vector of index_vec that it needs to reference.  Apparently when the
constructor checks if it is ready by testing its size, it sees a value
that indicates that it is not ready, most likely gets a return of 0, but
when the program exits, the ivs vector is destroyed before the plbck_strg
variable, and when the plbck_strg destructor tries to access the index_vec
that it was assigned, the Seg fault happens.  This problem is fixed by
adding an init_priority attribute to the definition of both globals.

I also added a build script bld-test-repo to build the test program.  It
shouldn't matter what the main name of the repository is, since as long as
the command is started in any of the repository folders, it should work.